fix: edit project architecture
