--- src/main/java/com/aloc/aloc/domain/history/History.java
@@ -1,35 +1,0 @@
-package com.aloc.aloc.domain.history;
-
-import com.aloc.aloc.domain.user.User;
-import com.aloc.aloc.global.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class History extends AuditingTimeEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Column(nullable = false)
-	private String username;
-
-	@Column(nullable = false)
-	private String icon;
-
-	private Integer rank;
-}
