--- src/main/java/com/aloc/aloc/domain/problemtag/ProblemTag.java
@@ -1,36 +1,0 @@
-package com.aloc.aloc.domain.problemtag;
-
-import com.aloc.aloc.domain.problem.Problem;
-import com.aloc.aloc.domain.tag.Tag;
-import com.aloc.aloc.global.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProblemTag extends AuditingTimeEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne
-	@JoinColumn(name = "problem_id", nullable = false)
-	private Problem problem;
-
-	@ManyToOne
-	@JoinColumn(name = "tag_id", nullable = false)
-	private Tag tag;
-}
