--- src/main/java/com/aloc/aloc/domain/problemtype/ProblemType.java
@@ -1,37 +1,0 @@
-package com.aloc.aloc.domain.problemtype;
-
-import com.aloc.aloc.domain.problemtype.enums.Course;
-import com.aloc.aloc.domain.problemtype.enums.Routine;
-import com.aloc.aloc.global.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProblemType extends AuditingTimeEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(nullable = false)
-	@Enumerated(EnumType.STRING)
-	private Routine routine;
-
-	@Column(nullable = false)
-	@Enumerated(EnumType.STRING)
-	private Course course;
-}
