--- src/main/java/com/aloc/aloc/domain/problemtype/enums/Course.java
@@ -1,5 +1,0 @@
-package com.aloc.aloc.domain.problemtype.enums;
-
-public enum Course {
-	HALF, FULL
-}
