--- src/main/java/com/aloc/aloc/domain/problemtype/enums/Routine.java
@@ -1,5 +1,0 @@
-package com.aloc.aloc.domain.problemtype.enums;
-
-public enum Routine {
-	WEEKLY, DAILY
-}
