--- src/main/java/com/aloc/aloc/domain/user/enums/Authority.java
@@ -1,5 +1,0 @@
-package com.aloc.aloc.domain.user.enums;
-
-public enum Authority {
-	ROLE_USER, ROLE_ADMIN, ROLE_GUEST
-}
