+++ src/main/java/com/aloc/aloc/global/jwt/service/JwtServiceImpl.java
@@ -9,8 +9,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.aloc.aloc.domain.user.User;
-import com.aloc.aloc.domain.user.repository.UserRepository;
+import com.aloc.aloc.user.User;
+import com.aloc.aloc.user.repository.UserRepository;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
