+++ src/main/java/com/aloc/aloc/global/login/service/UserDetailsServiceImpl.java
@@ -9,8 +9,8 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.aloc.aloc.domain.user.User;
-import com.aloc.aloc.domain.user.repository.UserRepository;
+import com.aloc.aloc.user.User;
+import com.aloc.aloc.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
