+++ src/main/java/com/aloc/aloc/problemtype/enums/Course.java
@@ -1,0 +1,5 @@
+package com.aloc.aloc.problemtype.enums;
+
+public enum Course {
+	HALF, FULL
+}
