+++ src/main/java/com/aloc/aloc/problemtype/enums/Routine.java
@@ -1,0 +1,5 @@
+package com.aloc.aloc.problemtype.enums;
+
+public enum Routine {
+	WEEKLY, DAILY
+}
