+++ src/main/java/com/aloc/aloc/solvedproblem/repository/ProblemSolvingCountProjection.java
@@ -1,0 +1,7 @@
+package com.aloc.aloc.solvedproblem.repository;
+
+public interface ProblemSolvingCountProjection {
+	Long getProblemId();
+
+	Integer getSolvingCount();
+}
