+++ src/main/java/com/aloc/aloc/tag/dto/TagSimpleDto.java
@@ -1,0 +1,16 @@
+package com.aloc.aloc.tag.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class TagSimpleDto {
+	private Long id;
+	private String koreanName;
+	private String englishName;
+}
