+++ src/main/java/com/aloc/aloc/user/dto/response/UserListResponseDto.java
@@ -1,0 +1,12 @@
+package com.aloc.aloc.user.dto.response;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserListResponseDto {
+	List<UserResponseDto> userList;
+}
