+++ src/main/java/com/aloc/aloc/user/enums/Authority.java
@@ -1,0 +1,5 @@
+package com.aloc.aloc.user.enums;
+
+public enum Authority {
+	ROLE_USER, ROLE_ADMIN, ROLE_GUEST
+}
